+++ turtlebot3_navigation2/launch/navigation2.launch.py
@@ -38,7 +38,7 @@ def generate_launch_description():
 
     param_file_name = TURTLEBOT3_MODEL + '.yaml'
     param_dir = LaunchConfiguration(
-        'params',
+        'params_file',
         default=os.path.join(
             get_package_share_directory('turtlebot3_navigation2'),
             'param',
@@ -58,7 +58,7 @@ def generate_launch_description():
             description='Full path to map file to load'),
 
         DeclareLaunchArgument(
-            'params',
+            'params_file',
             default_value=param_dir,
             description='Full path to param file to load'),
 
@@ -72,13 +72,13 @@ def generate_launch_description():
             launch_arguments={
                 'map': map_dir,
                 'use_sim_time': use_sim_time,
-                'params': param_dir}.items(),
+                'params_file': param_dir}.items(),
         ),
 
         Node(
             package='rviz2',
-            node_executable='rviz2',
-            node_name='rviz2',
+            executable='rviz2',
+            name='rviz2',
             arguments=['-d', rviz_config_dir],
             parameters=[{'use_sim_time': use_sim_time}],
             output='screen'),