+++ turtlebot3_node/include/turtlebot3_node/sensors/sensor_state.hpp
@@ -38,9 +38,9 @@ public:
     const std::string & topic_name = "sensor_state",
     const bool & bumper_forward = false,
     const bool & bumper_backward = false,
+    const bool & illumination = false,
     const bool & cliff = false,
-    const bool & sonar = false,
-    const bool & illumination = false);
+    const bool & sonar = false);
 
   void publish(
     const rclcpp::Time & now,
@@ -51,9 +51,9 @@ private:
 
   bool bumper_forward_;
   bool bumper_backward_;
+  bool illumination_;
   bool cliff_;
   bool sonar_;
-  bool illumination_;
 };
 }  // namespace sensors
 }  // namespace turtlebot3