+++ turtlebot3_example/CHANGELOG.rst
@@ -2,6 +2,10 @@
 Changelog for package turtlebot3_example
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+2.1.2 (2021-04-07)
+------------------
+* None
+
 2.1.1 (2021-01-06)
 ------------------
 * Nav2 prefix filename removed